initial breakout of AWS bits
